+++ bin/gitm-approve.ts
@@ -51,11 +51,12 @@ options.forEach((o: GitmarsOptionOptionsType) => {
 // .option('--state [state]', '筛选合并请求状态，共有2种：opened、closed，不传则默认全部', null)
 // .option('--quiet', '不要推送消息', false)
 program.action(async (opt: GitmBuildOption): Promise<void> => {
+    const userInfoApi = (config.apis && config.apis.userInfo) || config.api
     const {
         token,
         level,
         nickname = ''
-    } = config.api ? await getUserToken() : ({} as FetchDataType)
+    } = userInfoApi ? await getUserToken() : ({} as FetchDataType)
     if (level && level > 3) {
         echo(red(`${nickname}同学，你的权限不足`))
         process.exit(1)